+++ src/integrationTest/java/uk/gov/hmcts/reform/bulkscan/orchestrator/controllers/RootControllerTest.java
@@ -8,7 +8,7 @@ import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.CallbackProcessorImpl;
+import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.CallbackProcessor;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -16,7 +16,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @RunWith(SpringRunner.class)
 @WebMvcTest()
 @ActiveProfiles("integration")
-@Import({CallbackProcessorImpl.class})
+@Import({CallbackProcessor.class})
 public class RootControllerTest {
 
     @Autowired
